pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cpppo>=5.2.5",
    "flask>=3.1.2",
    "paho-mqtt>=2.1.0",
    "pycomm3>=1.2.14",
    "python-dotenv>=1.2.1",
]
